Conjunction search for genres
Allows to search for movies that has ALL required genres, not ONE of
required. Without this patch, it wasn't possible (for example) to find
"historical drama". Because if you tried to search by selecting History
and Drama, search would return all History (educational, history comedy,
...) and all Drama (sci-fi drama, ...). With this patch, search will
return movies that is History AND Drama.